+++ src/net/channel/error.rs
@@ -3,5 +3,5 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum ChannelError {
     #[error("Requested channel was not found in channel layer: {0}")]
-    NotFound(i8),
+    NotFound(i16),
 }
