+++ src/net/world/error.rs
@@ -3,5 +3,5 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum WorldError {
     #[error("Requested world was not found in world layer: {0}")]
-    NotFound(i8),
+    NotFound(i16),
 }
